+++ packages/core-infrastructure/frontend/src/index.ts
@@ -2,18 +2,18 @@ export {
 	createActorWithConfig,
 	loadConfig,
 	loadMockBackendFromModules,
-} from "./config";
-export { useActor } from "./hooks/useActor";
+} from "./config.js";
+export { useActor } from "./hooks/useActor.js";
 export {
 	type InternetIdentityContext,
 	InternetIdentityProvider,
 	type LoginOptions,
 	type Status,
 	useInternetIdentity,
-} from "./hooks/useInternetIdentity";
+} from "./hooks/useInternetIdentity.js";
 export type {
 	CreateActorOptions,
 	createActorFunction,
 	MockBackendOptions,
 	MockModules,
-} from "./types";
+} from "./types.js";